+++ fvtest/compilertest/tests/FooBarTest.cpp
@@ -25,8 +25,8 @@
 #include "ilgen/TypeDictionary.hpp"
 #include "OpCodesTest.hpp"
 #include "tests/FooBarTest.hpp"
-#include "tests/BarIlInjector.hpp"
-#include "tests/FooIlInjector.hpp"
+#include "tests/injectors/BarIlInjector.hpp"
+#include "tests/injectors/FooIlInjector.hpp"
 #include "gtest/gtest.h"
 #include "ilgen/IlGeneratorMethodDetails_inlines.hpp"
