+++ fvtest/compilertest/tests/OpCodesTest.cpp
@@ -30,17 +30,17 @@
 #include "env/jittypes.h"
 #include "gtest/gtest.h"
 #include "il/DataTypes.hpp"
-#include "ilgen/BinaryOpIlInjector.hpp"
-#include "ilgen/ChildlessUnaryOpIlInjector.hpp"
-#include "ilgen/CmpBranchOpIlInjector.hpp"
-#include "ilgen/IlGeneratorMethodDetails_inlines.hpp"
-#include "ilgen/StoreOpIlInjector.hpp"
-#include "ilgen/TernaryOpIlInjector.hpp"
 #include "ilgen/TypeDictionary.hpp"
-#include "ilgen/UnaryOpIlInjector.hpp"
-#include "tests/CallIlInjector.hpp"
-#include "tests/IndirectLoadIlInjector.hpp"
-#include "tests/IndirectStoreIlInjector.hpp"
+#include "ilgen/IlGeneratorMethodDetails_inlines.hpp"
+#include "tests/injectors/BinaryOpIlInjector.hpp"
+#include "tests/injectors/CallIlInjector.hpp"
+#include "tests/injectors/ChildlessUnaryOpIlInjector.hpp"
+#include "tests/injectors/CmpBranchOpIlInjector.hpp"
+#include "tests/injectors/IndirectLoadIlInjector.hpp"
+#include "tests/injectors/IndirectStoreIlInjector.hpp"
+#include "tests/injectors/StoreOpIlInjector.hpp"
+#include "tests/injectors/TernaryOpIlInjector.hpp"
+#include "tests/injectors/UnaryOpIlInjector.hpp"
 
 #include "tests/OpCodesTest.hpp"
