+++ fvtest/compilertest/tests/injectors/CmpBranchOpIlInjector.cpp
@@ -22,19 +22,32 @@
 #include "compile/Compilation.hpp"
 #include "env/FrontEnd.hpp"
 #include "compile/Method.hpp"
-#include "ilgen/UnaryOpIlInjector.hpp"
+#include "tests/injectors/CmpBranchOpIlInjector.hpp"
 
 namespace TestCompiler
 {
-
 bool
-UnaryOpIlInjector::injectIL()
+CmpBranchOpIlInjector::injectIL()
    {
    if (!isOpCodeSupported())
       return false;
+   createBlocks(3);
 
-   createBlocks(1);
-   returnValue(createWithoutSymRef(_opCode, 1, parm(1)));
+   // 3 blocks requested start at 2 (0 is entry, 1 is exit)
+   // by default, generate to block 2
+
+   // Block2: blocks(0)
+   // if () goto Block4;
+   ifjump(_opCode, parm(1), parm(2), 2);
+
+   // Block3: blocks(1)
+   // return 0;
+   returnValue(iconst(0));
+
+   // Block4: blocks(2)
+   // return 1;
+   generateToBlock(2);
+   returnValue(iconst(1));
 
    return true;
    }
